requirements.txt
ultralytics==8.1.22
torch==2.1.0
pillow==10.2.0
numpy==1.24.4
opencv-python==4.9.0.80
dvc[s3]==3.40.0
mlflow==2.10.1
pyyaml==6.0.1
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
